hsi-11591 Error al editar datos de la referencia

+++ front-end/apps/projects/hospital/src/app/modules/access-management/dialogs/reference-edition-pop-up/reference-edition-pop-up.component.ts
@@ -46,19 +46,25 @@ export class ReferenceEditionPopUpComponent implements OnInit {
 				this.newReferenceInfo = { ...this.newReferenceInfo, fileIds: this.referenceFiles.oldFiles.map(files => files.fileId) };
 			
 			if (this.data.isGestor) {
+				if (this.deletedIdsFiles.length)
 				this.referenceFileService.deleteReferenceFiles(this.deletedIdsFiles).pipe(take(1))
 				.subscribe(deleted => {
 						if (deleted) this.modifyReferenceAsGestor();
 					});
+
+				else this.modifyReferenceAsGestor();
 			}
 			else this.modifyReference();
 			return;
 		}
 		if (this.data.isGestor) {
+			if (this.deletedIdsFiles.length)
 			this.referenceFileService.deleteReferenceFiles(this.deletedIdsFiles).pipe(take(1))
 				.subscribe(deleted => {
 					if (deleted) this.loadNewFilesAndModifyReferenceAsGestor();
 				});
+
+			else this.loadNewFilesAndModifyReferenceAsGestor();
 		}
 		else this.loadFilesAndModifyReference();
 	}
